optimized apply_opacity function
- wip: refactoring floor caster

+++ include/graphics.h
@@ -6,7 +6,7 @@
 /*   By: kkhai-ki <kkhai-ki@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/23 10:00:47 by we                #+#    #+#             */
-/*   Updated: 2025/02/25 11:11:58 by kkhai-ki         ###   ########.fr       */
+/*   Updated: 2025/02/28 21:22:32 by kkhai-ki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 # include <math.h>
 # include "map.h"
 
-int				apply_opacity(int color, float opacity);
+int				apply_opacity(int color, int opacity);
 void			put_pixel(int x, int y, int color, t_img *img);
 void			draw_textured_column(int x, t_ray *ray, t_tex_data *tex_data, t_vars *vars);
 void			get_textures(int x, t_ray *ray, t_player *player, t_vars *mlx);
